// src/lib/types.ts
export type InquiryType = 
  | "Product & Purchasing Inquiry"
  | "Distribution Partnership Discussion"
  | "ODM/OEM/SKD Partnership Discussion"
  | "Other Agenda";

export type ProductInterest = 
  | "Monitor Defibrillators"
  | "Patient Monitors"
  | "AEDs"
  | "Others";

export interface Booking {
  id: string;
  name: string;
  email: string;
  companyName: string;
  country: string;
  productInterest: ProductInterest;
  inquiryType: InquiryType;
  message?: string;
  date: string; // YYYY-MM-DD
  time: string; // HH:mm
  createdAt: string;
}